requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==9.5.0
tensorflow==2.19.0
python-dotenv==1.1.1